Change ExperimentDetails to ActivityDetails and point to activities state
