+++ ui/src/components/activity/ActivityDetails/ActivityDetails.module.css
